+++ core/src/watchosMain/kotlin/com/powersync/DatabaseDriverFactory.watchos.kt
@@ -31,3 +31,10 @@ private val didLoadExtension by lazy {
 
     true
 }
+
+@ExperimentalPowerSyncAPI
+@Throws(PowerSyncException::class)
+public actual fun resolvePowerSyncLoadableExtensionPath(): String? {
+    didLoadExtension
+    return null
+}